feat: Add release workflows for each target with optional builds
