--- .github/workflows/publish-crates-eclipse.yml
@@ -1,38 +1,0 @@
-name: Publish crates (Eclipse)
-
-on:
-  workflow_call:
-    inputs:
-      live-run:
-        type: string
-        required: true
-      version:
-        type: string
-        required: true
-      name:
-        type: string
-        required: true
-  workflow_dispatch:
-    inputs:
-      live-run:
-        type: string
-        required: true
-      version:
-        type: string
-        required: true
-      name:
-        type: string
-        required: true
-
-jobs:
-  publish-crates-debian:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: eclipse-zenoh/ci/publish-crates-eclipse@main
-        with:
-          live-run: ${{ inputs.live-run }}
-          version: ${{ inputs.version }}
-          ssh-host: genie.zenoh@projects-storage.eclipse.org
-          ssh-host-path: /home/data/httpd/download.eclipse.org/${{ inputs.name }}
-          ssh-private-key: ${{ secrets.SSH_PRIVATE_KEY }}
-          ssh-passphrase: ${{ secrets.SSH_PASSPHRASE }}
