--- .github/workflows/publish-crates-homebrew.yml
@@ -1,65 +1,0 @@
-name: Publish crates (Homebrew)
-
-on:
-  workflow_call:
-    inputs:
-      repo:
-        type: string
-        required: true
-      live-run:
-        type: string
-        required: true
-      version:
-        type: string
-        required: true
-      branch:
-        type: string
-        required: true
-      artifact-patterns:
-        type: string
-        required: true
-      formulae:
-        type: string
-        required: true
-  workflow_dispatch:
-    inputs:
-      repo:
-        type: string
-        required: true
-      live-run:
-        type: string
-        required: true
-      version:
-        type: string
-        required: true
-      branch:
-        type: string
-        required: true
-      artifact-patterns:
-        type: string
-        required: true
-      formulae:
-        type: string
-        required: true
-
-jobs:
-  publish-crates-homebrew:
-    runs-on: macos-latest
-    steps:
-      - uses: eclipse-zenoh/ci/publish-crates-homebrew@main
-        with:
-          repo: ${{ inputs.repo }}
-          live-run: ${{ inputs.live-run }}
-          version: ${{ inputs.version }}
-          branch: ${{ inputs.branch }}
-          formulae: ${{ inputs.formulae }}
-          artifact-patterns: ${{ inputs.artifact-patterns }}
-          tap: eclipse-zenoh/homebrew-zenoh
-          ssh-host: genie.zenoh@projects-storage.eclipse.org
-          ssh-host-path: /home/data/httpd/download.eclipse.org/zenoh/homebrew-tap
-          ssh-host-url: https://download.eclipse.org/zenoh/homebrew-tap
-          actor-name: eclipse-zenoh-bot
-          actor-email: eclipse-zenoh-bot@users.noreply.github.com
-          ssh-private-key: ${{ secrets.SSH_PRIVATE_KEY }}
-          ssh-passphrase: ${{ secrets.SSH_PASSPHRASE }}
-          github-token: ${{ secrets.BOT_TOKEN_WORKFLOW }}
