+++ .github/workflows/release-crates-cargo.yml
@@ -1,0 +1,44 @@
+name: Publish crates (Cargo)
+
+on:
+  workflow_call:
+    inputs:
+      repos:
+        type: string
+        required: true
+      live-run:
+        type: string
+        required: true
+      branch:
+        type: string
+        required: true
+      inter-deps-pattern:
+        type: string
+        required: true
+  workflow_dispatch:
+    inputs:
+      repos:
+        type: string
+        required: true
+      live-run:
+        type: string
+        required: true
+      branch:
+        type: string
+        required: true
+      inter-deps-pattern:
+        type: string
+        required: true
+
+jobs:
+  publish:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: eclipse-zenoh/ci/publish-crates-cargo@main
+        with:
+          repos: ${{ inputs.repos }}
+          live-run: ${{ inputs.live-run }}
+          branch: ${{ inputs.branch }}
+          inter-deps-pattern: ${{ inputs.inter-deps-pattern }}
+          github-token: ${{ secrets.BOT_TOKEN_WORKFLOW }}
+          crates-io-token: ${{ secrets.CRATES_IO_TOKEN }}
