+++ .github/workflows/release-crates-debian.yml
@@ -1,0 +1,61 @@
+name: Publish crates (Debian)
+
+on:
+  workflow_call:
+    inputs:
+      live-run:
+        type: string
+        required: true
+      repo:
+        type: string
+        required: true
+      version:
+        type: string
+        required: true
+      branch:
+        type: string
+        required: true
+      no-build:
+        type: boolean
+        required: false
+        default: false
+  workflow_dispatch:
+    inputs:
+      live-run:
+        type: string
+        required: true
+      repo:
+        type: string
+        required: true
+      version:
+        type: string
+        required: true
+      branch:
+        type: string
+        required: true
+      no-build:
+        type: boolean
+        required: false
+        default: false
+
+jobs:
+  build:
+    if: ${{ !inputs.no-build }}
+    uses: eclipse-zenoh/ci/.github/workflows/build-crates-debian.yml@main
+    with:
+      repo: ${{ inputs.repo }}
+      version: ${{ inputs.version }}
+      branch: ${{ inputs.branch }}
+
+  publish:
+    needs: build
+    runs-on: ubuntu-latest
+    steps:
+      - uses: eclipse-zenoh/ci/publish-crates-debian@main
+        with:
+          live-run: ${{ inputs.live-run }}
+          version: ${{ inputs.version }}
+          ssh-host: genie.zenoh@projects-storage.eclipse.org
+          ssh-host-path: /home/data/httpd/download.eclipse.org/zenoh/debian-repo
+          ssh-private-key: ${{ secrets.SSH_PRIVATE_KEY }}
+          ssh-passphrase: ${{ secrets.SSH_PASSPHRASE }}
