--- build-crates-artifacts/action.yml
@@ -1,23 +1,0 @@
-name: Build crates (Artifacts)
-
-inputs:
-  repo:
-    required: true
-  version:
-    required: true
-  branch:
-    required: true
-  target:
-    required: true
-  artifact-patterns:
-    required: true
-  github-token:
-    required: true
-
-outputs:
-  artifact-id:
-    description: Artifact id
-
-runs:
-  using: node20
-  main: ../dist/build-crates-artifacts-main.js
