+++ package.json
@@ -1,6 +1,6 @@
 {
-  "name": "zenoh-actions",
-  "description": "Actions for Eclipse Zenoh CI/CD",
+  "name": "@eclipse-zenoh/ci",
+  "description": "GitHub JavaScript Actions for Eclipse Zenoh CI/CD",
   "private": true,
   "engines": {
     "node": ">=20"
@@ -15,7 +15,7 @@
     "all": "npm run format:write && npm run lint && npm run test && npm run build",
     "clean": "rm dist/*"
   },
-  "license": "Apache-2.0",
+  "license": "(Apache-2.0 OR EPL-2.0)",
   "dependencies": {
     "@actions/artifact": "^2.1.1",
     "@actions/cache": "^3.2.4",
