--- src/build-crates-artifacts-main.ts
@@ -1,3 +1,0 @@
-import { main, setup } from "./build-crates-artifacts";
-
-await main(setup());
