+++ src/build-crates-standalone-main.ts
@@ -1,0 +1,3 @@
+import { main, setup } from "./build-crates-standalone";
+
+await main(setup());
