+++ src/publish-crates-homebrew.ts
@@ -2,18 +2,19 @@ import * as fs from "fs/promises";
 import * as crypto from "crypto";
 
 import * as core from "@actions/core";
+import { DefaultArtifactClient } from "@actions/artifact";
 
 import { sh } from "./command";
-import * as path from "path";
 import * as ssh from "./ssh";
-import * as zip from "./zip";
+
+import { artifactName } from "./build-crates-standalone";
+
+const artifact = new DefaultArtifactClient();
 
 export type Input = {
   liveRun: boolean;
   version: string;
   repo: string;
-  branch: string;
-  artifactRegExp: RegExp;
   formulae: string[];
   tap: string;
   sshHost: string;
@@ -29,8 +30,6 @@ export function setup(): Input {
   const liveRun = core.getInput("live-run");
   const version = core.getInput("version", { required: true });
   const repo = core.getInput("repo", { required: true });
-  const branch = core.getInput("branch", { required: true });
-  const artifactPatterns = core.getInput("artifact-patterns", { required: true });
   const formulae = core.getInput("formulae", { required: true });
   const tap = core.getInput("tap", { required: true });
   const sshHost = core.getInput("ssh-host", { required: true });
@@ -46,8 +45,6 @@ export function setup(): Input {
     liveRun: liveRun == "" ? false : core.getBooleanInput("live-run"),
     version,
     repo,
-    branch,
-    artifactRegExp: new RegExp(artifactPatterns.split("\n").join("|")),
     formulae: formulae.split("\n"),
     tap,
     sshHost,
@@ -68,8 +65,6 @@ export function setup(): Input {
 const X86_64_APPLE_DARWIN = "x86_64-apple-darwin";
 const AARCH64_APPLE_DARWIN = "aarch64-apple-darwin";
 
-const TARGETS = [X86_64_APPLE_DARWIN, AARCH64_APPLE_DARWIN];
-
 const AARCH64_URL = "aarch64-url";
 const AARCH64_SHA256 = "aarch64-sha256";
 const X86_64_URL = "x86_64-url";
@@ -87,27 +82,18 @@ type Release = {
 export async function main(input: Input) {
   try {
     const repo = input.repo.split("/").at(1);
-    const remote = `https://${input.githubToken}@github.com/${input.repo}.git`;
-    const tapPath = `${sh(`brew --repository`).trim()}/Library/Taps/${input.tap}`;
+    const tapPath = `${sh("brew --repository").trim()}/Library/Taps/${input.tap}`;
     const tapUrl = `https://${input.githubToken}@github.com/${input.tap}.git`;
 
-    sh(`git clone --recursive --single-branch --branch ${input.branch} ${remote}`);
-
-    for (const target of TARGETS) {
-      sh(`rustup target add ${target}`, { cwd: repo });
-    }
-
-    // NOTE: See Cargo's (-Z) multitarget feature
-    sh(`cargo build --release --bins --lib ${TARGETS.map(t => `--target ${t}`).join(" ")}`, { cwd: repo });
-
-    for (const target of TARGETS) {
-      const outputArchive = `${repo}-${input.version}-${target}.zip`;
-      await zip.fromDirectory(outputArchive, path.join(repo, "target", target, "release"), input.artifactRegExp);
+    for (const target of [X86_64_APPLE_DARWIN, AARCH64_APPLE_DARWIN]) {
+      const name = artifactName(repo, input.version, target);
+      const result = await artifact.getArtifact(name);
+      await artifact.downloadArtifact(result.artifact.id);
 
       if (input.liveRun) {
         await ssh.withIdentity(input.sshPrivateKey, input.sshPassphrase, env => {
           sh(`ssh -v -o StrictHostKeyChecking=no ${input.sshHost} mkdir -p ${input.sshHostPath}`, { env });
-          sh(`scp -v -o StrictHostKeyChecking=no -r ${outputArchive} ${input.sshHost}:${input.sshHostPath}`, { env });
+          sh(`scp -v -o StrictHostKeyChecking=no -r ${name} ${input.sshHost}:${input.sshHostPath}`, { env });
         });
       }
     }
@@ -120,14 +106,13 @@ export async function main(input: Input) {
     const release = JSON.parse(releaseFile) as Release;
 
     const sha256 = async (target: string): Promise<string> => {
-      const archive = `${repo}-${input.version}-${target}.zip`;
-      const contents = await fs.readFile(archive);
+      const contents = await fs.readFile(artifactName(repo, input.version, target));
       return crypto.createHash("sha256").update(contents).digest("hex");
     };
 
     const url = (target: string): string => {
       const baseUrl = input.liveRun ? input.sshHostUrl : `file://${process.cwd()}`;
-      return `${baseUrl}/${repo}-${input.version}-${target}.zip`;
+      return `${baseUrl}/${artifactName(repo, input.version, target)}`;
     };
 
     for (const formula of input.formulae) {
@@ -147,6 +132,7 @@ export async function main(input: Input) {
       sh(`brew audit ${formula}`);
       sh(`brew install --force ${formula}`);
       sh(`brew uninstall --force --ignore-dependencies ${formula}`);
+      sh("brew autoremove");
     }
 
     if (input.liveRun) {
